models: move _git
